backend/requirements.txt
asgiref==3.11.1
channels==4.3.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.1.0
sqlparse==0.5.5
tzdata==2025.3
django-environ==0.13.0